Updated tictactoe logic tests, updated game logic
Updated tictactoe position validity checks to also include checks for if a player makes a move after the other player has won the game.

Updated tic tac toe game to count moves and keep track of the current player at the initialization of the board and when a cell in the board is updated. (untested)

Updated game logic to make game function (tested [not extensively], appears to be functional)